Rename metadata classes/package. Update calibrate_y_factor description.
