+++ scos_actions/actions/acquire_single_freq_tdomain_iq.py
@@ -40,10 +40,10 @@ from scos_actions.actions.interfaces.action import Action
 from scos_actions.actions.interfaces.signals import measurement_action_completed
 from scos_actions.actions.sigmf_builder import Domain, MeasurementType, SigMFBuilder
 from scos_actions.hardware import gps as mock_gps
-from scos_actions.actions.metadata_decorators.time_domain_annotation import TimeDomainAnnotation
-from scos_actions.actions.metadata_decorators.calibration_annotation import CalibrationAnnotation
-from scos_actions.actions.metadata_decorators.measurement_global import MeasurementMetadata
-from scos_actions.actions.metadata_decorators.sensor_annotation import SensorAnnotation
+from scos_actions.actions.metadata.time_domain_annotation import TimeDomainAnnotation
+from scos_actions.actions.metadata.calibration_annotation import CalibrationAnnotation
+from scos_actions.actions.metadata.measurement_global import MeasurementMetadata
+from scos_actions.actions.metadata.sensor_annotation import SensorAnnotation
 
 logger = logging.getLogger(__name__)
